Updates on ephemeris calculations, location, etc.
For issue #1, update to using ephem for accuracy (and geekery) as well
as use google API to retrieve a lat/long pair based on location setting.
You can get accurate with an address, or just use a city and the Google
geocode API will return lat/long values.

Pickle the data we get from the geocode API so we don't make further
requests once we've saved that file off.

TODO:
Figure out use of the pickle for other settings we might want to cache
long-term per-user.